refactor(history/list.rs,history/table.rs): Renamed list.rs to table.rs and refactored it in order to pass a single params signal
